feat: Implement TokenValidator component
- Added `TokenValidator` class in `src/coreason_identity/token_validator.py`.
- Implemented JWT signature and claim validation using `authlib`.
- Enforced strict checks for `iss`, `aud`, and `exp`.
- Mapped external exceptions to package-specific `TokenExpiredError`, `InvalidAudienceError`, and `SignatureVerificationError`.
- Added comprehensive unit tests in `tests/test_token_validator.py` with 100% coverage.
- Updated `.pre-commit-config.yaml` to include `pytest` in `mypy` dependencies for test type checking.
- Updated `OIDCProvider` to accept `float` for `cache_ttl` to improve testing compatibility.
- Verified type safety with `mypy` and code style with `ruff`.
- Ensured no regressions in existing tests.

+++ src/coreason_identity/oidc_provider.py
@@ -25,7 +25,7 @@ class OIDCProvider:
     Fetches and caches the Identity Provider's configuration and JWKS.
     """
 
-    def __init__(self, discovery_url: str, cache_ttl: int = 3600) -> None:
+    def __init__(self, discovery_url: str, cache_ttl: float = 3600) -> None:
         """
         Initialize the OIDCProvider.
